requirements.txt
Django==3.1.4
pipenv==2023.12.1
robotframework==7.0
robotframework-pythonlibcore==4.4.1
robotframework-seleniumlibrary==6.2.0
selenium==4.19.0
chromedriver-autoinstaller==0.2.2